+++ src/main/java/DarrenBot/exception/EmptyTaskException.java
@@ -1,0 +1,7 @@
+package DarrenBot.exception;
+
+public class EmptyTaskException extends Exception {
+    public EmptyTaskException(String taskType) {
+        super("OOPS!!! The description of a " + taskType + " cannot be empty.");
+    }
+}
