+++ src/main/java/DarrenBot/exception/UnexpectedCommandException.java
@@ -1,0 +1,7 @@
+package DarrenBot.exception;
+
+public class UnexpectedCommandException extends Exception {
+    public UnexpectedCommandException(String message) {
+        super(message);
+    }
+}
