+++ src/main/java/DarrenBot/tasks/Task.java
@@ -1,0 +1,29 @@
+package DarrenBot.tasks;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public void undo() {
+        this.isDone = false;
+    }
+
+    public void redo() {
+        this.isDone = true;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    @Override
+    public String toString() {
+        return "[" + this.getStatusIcon() + "] " + this.description;
+    }
+}
+
