+++ src/main/java/DarrenBot/tasks/Todo.java
@@ -1,0 +1,14 @@
+package DarrenBot.tasks;
+
+public class Todo extends Task {
+    String taskType;
+    public Todo(String description) {
+        super(description);
+        this.taskType = "todo";
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
